pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "omnigen"
version = "0.1.0"
description = "OmniGen"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "openai>=1.97.0",
    "python-dotenv>=1.1.1",
]

[project.optional-dependencies]
dev = ["pandas>=2.3.1", "pyarrow>=21.0.0", "pyyaml>=6.0.2"]
test = ["pytest>=8.4.1", "pytest-asyncio>=1.1.0"]

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.hatch.metadata]
allow-direct-references = true

[tool.black]
line-length = 120
skip-string-normalization = true
